refactor: GameMachine 제거 및 Controller 네이밍 변경
