+++ src/main/java/blackjack/Application.java
@@ -1,12 +1,10 @@
 package blackjack;
 
-import blackjack.controller.BlackjackController;
-import blackjack.domain.GameMachine;
-import blackjack.view.InputView;
+import blackjack.controller.Blackjack;
 
 public class Application {
 
     public static void main(String[] args) {
-        new BlackjackController().play();
+        new Blackjack().play();
     }
 }
